Victory is mine. full java 9 Module Achieved

+++ org.springframework.petclinic.main/src/test/java/org/springframework/samples/petclinic/service/EntityUtils.java
@@ -41,7 +41,7 @@ public abstract class EntityUtils {
      * @return the found entity
      * @throws ObjectRetrievalFailureException if the entity was not found
      */
-    public static <T extends BaseEntity> T getById(Collection<T> entities, Class<T> entityClass, int entityId)
+    public static <T extends BaseEntity> T getById(Collection<T> entities, Class<T> entityClass, Long entityId)
         throws ObjectRetrievalFailureException {
         for (T entity : entities) {
             if (entity.getId() == entityId && entityClass.isInstance(entity)) {
